Rekey, and freshen build instructions for running on Ubuntu 18.04
Adds a few new dependencies to the apt-get command.

Replaces /bin/sh in shebang line with /bin/bash
(Ubuntu used to have /bin/sh point to bash, but now it points to dash:
  https://wiki.ubuntu.com/DashAsBinSh
The shell scripts use the "source" bash builtin which is missing from dash.)

Busybox apparently used to install in /usr/bin but now it's in /bin.